Aut 2717 fail fast

+++ acceptance-tests/src/test/java/uk/gov/di/test/step_definitions/Hooks.java
@@ -3,6 +3,7 @@ package uk.gov.di.test.step_definitions;
 import io.cucumber.java.After;
 import io.cucumber.java.AfterStep;
 import io.cucumber.java.Before;
+import io.cucumber.java.PendingException;
 import io.cucumber.java.Scenario;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
@@ -12,15 +13,27 @@ import java.net.MalformedURLException;
 
 public class Hooks extends BasePage {
 
-    @Before
+    private static int failureCount = 0;
+
+    protected static final Boolean FAIL_FAST_ENABLED =
+            Boolean.parseBoolean(System.getenv().getOrDefault("FAIL_FAST_ENABLED", "false"));
+
+    @Before(order = 2)
     public void setupWebdriver() throws MalformedURLException {
-        super.setupWebdriver();
-        driver.manage().deleteAllCookies();
+        if (failureCount == 0) {
+            super.setupWebdriver();
+            driver.manage().deleteAllCookies();
+        }
     }
 
-    @Before
+    @Before(order = 1)
     public void setUpScenario(Scenario scenario) {
         BasePage.scenario = scenario;
+        if (FAIL_FAST_ENABLED) {
+            if (failureCount > 0) {
+                throw new PendingException();
+            }
+        }
     }
 
     @AfterStep
@@ -35,6 +48,9 @@ public class Hooks extends BasePage {
             scenario.attach(screenshot, "image/png", "Failure screenshot");
 
             super.closeWebdriver();
+            if (FAIL_FAST_ENABLED) {
+                failureCount++;
+            }
         }
     }
 }
